ethdev: fix storage type of latest port id
eth_dev_last_created_port is used to store port id type and should
be extended to 16bits corresponding to ethdev port id range.

Fixes: f8244c6399d9 ("ethdev: increase port id range")
Cc: stable@dpdk.org

Acked-by: Ferruh Yigit <ferruh.yigit@intel.com>

+++ lib/librte_ethdev/rte_ethdev.c
@@ -49,7 +49,7 @@ static int ethdev_logtype;
 
 static const char *MZ_RTE_ETH_DEV_DATA = "rte_eth_dev_data";
 struct rte_eth_dev rte_eth_devices[RTE_MAX_ETHPORTS];
-static uint8_t eth_dev_last_created_port;
+static uint16_t eth_dev_last_created_port;
 
 /* spinlock for eth device callbacks */
 static rte_spinlock_t rte_eth_dev_cb_lock = RTE_SPINLOCK_INITIALIZER;
